@@ -18,19 +18,14 @@ import pytest
 
 from create_instance import main
 
-PROJECT = os.environ['GOOGLE_CLOUD_PROJECT']
-BUCKET = os.environ['CLOUD_STORAGE_BUCKET']
+PROJECT = os.environ["GOOGLE_CLOUD_PROJECT"]
+BUCKET = os.environ["CLOUD_STORAGE_BUCKET"]
 
 
 @pytest.mark.flaky(max_runs=3, min_passes=1)
 def test_main(capsys):
-    instance_name = 'test-instance-{}'.format(uuid.uuid4())
-    main(
-        PROJECT,
-        BUCKET,
-        'us-central1-f',
-        instance_name,
-        wait=False)
+    instance_name = "test-instance-{}".format(uuid.uuid4())
+    main(PROJECT, BUCKET, "us-central1-f", instance_name, wait=False)
 
     out, _ = capsys.readouterr()
 
