@@ -13,7 +13,7 @@
 # limitations under the License.
 
 import unittest
-import mock
+from unittest import mock
 
 import google.auth.credentials
 import google_auth_httplib2
@@ -89,7 +89,9 @@ class TestAuthWithGoogleAuth(unittest.TestCase):
 
         self.assertNotEqual(credentials, returned)
         self.assertEqual(returned, credentials.with_scopes.return_value)
-        credentials.with_scopes.assert_called_once_with(mock.sentinel.scopes, default_scopes=None)
+        credentials.with_scopes.assert_called_once_with(
+            mock.sentinel.scopes, default_scopes=None
+        )
 
     def test_authorized_http(self):
         credentials = mock.Mock(spec=google.auth.credentials.Credentials)
