@@ -1,13 +1,6 @@
 """Retain apiclient as an alias for googleapiclient."""
 
-import googleapiclient
-
-from googleapiclient import channel
-from googleapiclient import discovery
-from googleapiclient import errors
-from googleapiclient import http
-from googleapiclient import mimeparse
-from googleapiclient import model
+from googleapiclient import channel, discovery, errors, http, mimeparse, model
 
 try:
     from googleapiclient import sample_tools
