@@ -21,10 +21,11 @@ from __future__ import absolute_import
 __author__ = "jcgregorio@google.com (Joe Gregorio)"
 __all__ = ["build", "build_from_document", "fix_method_name", "key2param"]
 
-# Standard library imports
-import copy
 from collections import OrderedDict
 import collections.abc
+
+# Standard library imports
+import copy
 from email.generator import BytesGenerator
 from email.mime.multipart import MIMEMultipart
 from email.mime.nonmultipart import MIMENonMultipart
@@ -38,13 +39,14 @@ import os
 import re
 import urllib
 
+import google.api_core.client_options
+from google.auth.exceptions import MutualTLSChannelError
+from google.auth.transport import mtls
+from google.oauth2 import service_account
+
 # Third-party imports
 import httplib2
 import uritemplate
-import google.api_core.client_options
-from google.auth.transport import mtls
-from google.auth.exceptions import MutualTLSChannelError
-from google.oauth2 import service_account
 
 try:
     import google_auth_httplib2
@@ -52,29 +54,27 @@ except ImportError:  # pragma: NO COVER
     google_auth_httplib2 = None
 
 # Local imports
-from googleapiclient import _auth
-from googleapiclient import mimeparse
-from googleapiclient.errors import HttpError
-from googleapiclient.errors import InvalidJsonError
-from googleapiclient.errors import MediaUploadSizeError
-from googleapiclient.errors import UnacceptableMimeTypeError
-from googleapiclient.errors import UnknownApiNameOrVersion
-from googleapiclient.errors import UnknownFileType
-from googleapiclient.http import build_http
-from googleapiclient.http import BatchHttpRequest
-from googleapiclient.http import HttpMock
-from googleapiclient.http import HttpMockSequence
-from googleapiclient.http import HttpRequest
-from googleapiclient.http import MediaFileUpload
-from googleapiclient.http import MediaUpload
-from googleapiclient.model import JsonModel
-from googleapiclient.model import MediaModel
-from googleapiclient.model import RawModel
+from googleapiclient import _auth, mimeparse
+from googleapiclient._helpers import _add_query_parameter, positional
+from googleapiclient.errors import (
+    HttpError,
+    InvalidJsonError,
+    MediaUploadSizeError,
+    UnacceptableMimeTypeError,
+    UnknownApiNameOrVersion,
+    UnknownFileType,
+)
+from googleapiclient.http import (
+    BatchHttpRequest,
+    HttpMock,
+    HttpMockSequence,
+    HttpRequest,
+    MediaFileUpload,
+    MediaUpload,
+    build_http,
+)
+from googleapiclient.model import JsonModel, MediaModel, RawModel
 from googleapiclient.schema import Schemas
-
-from googleapiclient._helpers import _add_query_parameter
-from googleapiclient._helpers import positional
-
 
 # The client library requires a version of httplib2 that supports RETRIES.
 httplib2.RETRIES = 1
@@ -134,13 +134,13 @@ class _BytesGenerator(BytesGenerator):
 def fix_method_name(name):
     """Fix method names to avoid '$' characters and reserved word conflicts.
 
-  Args:
-    name: string, method name.
+    Args:
+      name: string, method name.
 
-  Returns:
-    The name with '_' appended if the name is a reserved word and '$' and '-'
-    replaced with '_'.
-  """
+    Returns:
+      The name with '_' appended if the name is a reserved word and '$' and '-'
+      replaced with '_'.
+    """
     name = name.replace("$", "_").replace("-", "_")
     if keyword.iskeyword(name) or name in RESERVED_WORDS:
         return name + "_"
@@ -151,14 +151,14 @@ def fix_method_name(name):
 def key2param(key):
     """Converts key names into parameter names.
 
-  For example, converting "max-results" -> "max_results"
+    For example, converting "max-results" -> "max_results"
 
-  Args:
-    key: string, the method key name.
+    Args:
+      key: string, the method key name.
 
-  Returns:
-    A safe method name based on the key name.
-  """
+    Returns:
+      A safe method name based on the key name.
+    """
     result = []
     key = list(key)
     if not key[0].isalpha():
@@ -193,72 +193,72 @@ def build(
 ):
     """Construct a Resource for interacting with an API.
 
-  Construct a Resource object for interacting with an API. The serviceName and
-  version are the names from the Discovery service.
+    Construct a Resource object for interacting with an API. The serviceName and
+    version are the names from the Discovery service.
 
-  Args:
-    serviceName: string, name of the service.
-    version: string, the version of the service.
-    http: httplib2.Http, An instance of httplib2.Http or something that acts
-      like it that HTTP requests will be made through.
-    discoveryServiceUrl: string, a URI Template that points to the location of
-      the discovery service. It should have two parameters {api} and
-      {apiVersion} that when filled in produce an absolute URI to the discovery
-      document for that service.
-    developerKey: string, key obtained from
-      https://code.google.com/apis/console.
-    model: googleapiclient.Model, converts to and from the wire format.
-    requestBuilder: googleapiclient.http.HttpRequest, encapsulator for an HTTP
-      request.
-    credentials: oauth2client.Credentials or
-      google.auth.credentials.Credentials, credentials to be used for
-      authentication.
-    cache_discovery: Boolean, whether or not to cache the discovery doc.
-    cache: googleapiclient.discovery_cache.base.CacheBase, an optional
-      cache object for the discovery documents.
-    client_options: Mapping object or google.api_core.client_options, client
-      options to set user options on the client.
-      (1) The API endpoint should be set through client_options. If API endpoint
-      is not set, `GOOGLE_API_USE_MTLS_ENDPOINT` environment variable can be used
-      to control which endpoint to use.
-      (2) client_cert_source is not supported, client cert should be provided using
-      client_encrypted_cert_source instead. In order to use the provided client
-      cert, `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be
-      set to `true`.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    adc_cert_path: str, client certificate file path to save the application
-      default client certificate for mTLS. This field is required if you want to
-      use the default client certificate. `GOOGLE_API_USE_CLIENT_CERTIFICATE`
-      environment variable must be set to `true` in order to use this field,
-      otherwise this field doesn't nothing.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    adc_key_path: str, client encrypted private key file path to save the
-      application default client encrypted private key for mTLS. This field is
-      required if you want to use the default client certificate.
-      `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be set to
-      `true` in order to use this field, otherwise this field doesn't nothing.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    num_retries: Integer, number of times to retry discovery with
-      randomized exponential backoff in case of intermittent/connection issues.
-    static_discovery: Boolean, whether or not to use the static discovery docs
-      included in the library. The default value for `static_discovery` depends
-      on the value of `discoveryServiceUrl`. `static_discovery` will default to
-      `True` when `discoveryServiceUrl` is also not provided, otherwise it will
-      default to `False`.
-    always_use_jwt_access: Boolean, whether always use self signed JWT for service
-      account credentials. This only applies to
-      google.oauth2.service_account.Credentials.
+    Args:
+      serviceName: string, name of the service.
+      version: string, the version of the service.
+      http: httplib2.Http, An instance of httplib2.Http or something that acts
+        like it that HTTP requests will be made through.
+      discoveryServiceUrl: string, a URI Template that points to the location of
+        the discovery service. It should have two parameters {api} and
+        {apiVersion} that when filled in produce an absolute URI to the discovery
+        document for that service.
+      developerKey: string, key obtained from
+        https://code.google.com/apis/console.
+      model: googleapiclient.Model, converts to and from the wire format.
+      requestBuilder: googleapiclient.http.HttpRequest, encapsulator for an HTTP
+        request.
+      credentials: oauth2client.Credentials or
+        google.auth.credentials.Credentials, credentials to be used for
+        authentication.
+      cache_discovery: Boolean, whether or not to cache the discovery doc.
+      cache: googleapiclient.discovery_cache.base.CacheBase, an optional
+        cache object for the discovery documents.
+      client_options: Mapping object or google.api_core.client_options, client
+        options to set user options on the client.
+        (1) The API endpoint should be set through client_options. If API endpoint
+        is not set, `GOOGLE_API_USE_MTLS_ENDPOINT` environment variable can be used
+        to control which endpoint to use.
+        (2) client_cert_source is not supported, client cert should be provided using
+        client_encrypted_cert_source instead. In order to use the provided client
+        cert, `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be
+        set to `true`.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      adc_cert_path: str, client certificate file path to save the application
+        default client certificate for mTLS. This field is required if you want to
+        use the default client certificate. `GOOGLE_API_USE_CLIENT_CERTIFICATE`
+        environment variable must be set to `true` in order to use this field,
+        otherwise this field doesn't nothing.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      adc_key_path: str, client encrypted private key file path to save the
+        application default client encrypted private key for mTLS. This field is
+        required if you want to use the default client certificate.
+        `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be set to
+        `true` in order to use this field, otherwise this field doesn't nothing.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      num_retries: Integer, number of times to retry discovery with
+        randomized exponential backoff in case of intermittent/connection issues.
+      static_discovery: Boolean, whether or not to use the static discovery docs
+        included in the library. The default value for `static_discovery` depends
+        on the value of `discoveryServiceUrl`. `static_discovery` will default to
+        `True` when `discoveryServiceUrl` is also not provided, otherwise it will
+        default to `False`.
+      always_use_jwt_access: Boolean, whether always use self signed JWT for service
+        account credentials. This only applies to
+        google.oauth2.service_account.Credentials.
 
-  Returns:
-    A Resource object with methods for interacting with the service.
+    Returns:
+      A Resource object with methods for interacting with the service.
 
-  Raises:
-    google.auth.exceptions.MutualTLSChannelError: if there are any problems
-      setting up mutual TLS channel.
-  """
+    Raises:
+      google.auth.exceptions.MutualTLSChannelError: if there are any problems
+        setting up mutual TLS channel.
+    """
     params = {"api": serviceName, "apiVersion": version}
 
     # The default value for `static_discovery` depends on the value of
@@ -328,16 +328,16 @@ def build(
 
 def _discovery_service_uri_options(discoveryServiceUrl, version):
     """
-    Returns Discovery URIs to be used for attempting to build the API Resource.
+      Returns Discovery URIs to be used for attempting to build the API Resource.
 
-  Args:
-    discoveryServiceUrl:
-        string, the Original Discovery Service URL preferred by the customer.
-    version:
-        string, API Version requested
+    Args:
+      discoveryServiceUrl:
+          string, the Original Discovery Service URL preferred by the customer.
+      version:
+          string, API Version requested
 
-  Returns:
-      A list of URIs to be tried for the Service Discovery, in order.
+    Returns:
+        A list of URIs to be tried for the Service Discovery, in order.
     """
 
     if discoveryServiceUrl is not None:
@@ -361,29 +361,29 @@ def _retrieve_discovery_doc(
     cache=None,
     developerKey=None,
     num_retries=1,
-    static_discovery=True
+    static_discovery=True,
 ):
     """Retrieves the discovery_doc from cache or the internet.
 
-  Args:
-    url: string, the URL of the discovery document.
-    http: httplib2.Http, An instance of httplib2.Http or something that acts
-      like it through which HTTP requests will be made.
-    cache_discovery: Boolean, whether or not to cache the discovery doc.
-    serviceName: string, name of the service.
-    version: string, the version of the service.
-    cache: googleapiclient.discovery_cache.base.Cache, an optional cache
-      object for the discovery documents.
-    developerKey: string, Key for controlling API usage, generated
-      from the API Console.
-    num_retries: Integer, number of times to retry discovery with
-      randomized exponential backoff in case of intermittent/connection issues.
-    static_discovery: Boolean, whether or not to use the static discovery docs
-      included in the library.
+    Args:
+      url: string, the URL of the discovery document.
+      http: httplib2.Http, An instance of httplib2.Http or something that acts
+        like it through which HTTP requests will be made.
+      cache_discovery: Boolean, whether or not to cache the discovery doc.
+      serviceName: string, name of the service.
+      version: string, the version of the service.
+      cache: googleapiclient.discovery_cache.base.Cache, an optional cache
+        object for the discovery documents.
+      developerKey: string, Key for controlling API usage, generated
+        from the API Console.
+      num_retries: Integer, number of times to retry discovery with
+        randomized exponential backoff in case of intermittent/connection issues.
+      static_discovery: Boolean, whether or not to use the static discovery docs
+        included in the library.
 
-  Returns:
-    A unicode string representation of the discovery document.
-  """
+    Returns:
+      A unicode string representation of the discovery document.
+    """
     from . import discovery_cache
 
     if cache_discovery:
@@ -401,7 +401,9 @@ def _retrieve_discovery_doc(
         if content:
             return content
         else:
-            raise UnknownApiNameOrVersion("name: %s  version: %s" % (serviceName, version))
+            raise UnknownApiNameOrVersion(
+                "name: %s  version: %s" % (serviceName, version)
+            )
 
     actual_url = url
     # REMOTE_ADDR is defined by the CGI spec [RFC3875] as the environment
@@ -451,63 +453,63 @@ def build_from_document(
 ):
     """Create a Resource for interacting with an API.
 
-  Same as `build()`, but constructs the Resource object from a discovery
-  document that is it given, as opposed to retrieving one over HTTP.
+    Same as `build()`, but constructs the Resource object from a discovery
+    document that is it given, as opposed to retrieving one over HTTP.
 
-  Args:
-    service: string or object, the JSON discovery document describing the API.
-      The value passed in may either be the JSON string or the deserialized
-      JSON.
-    base: string, base URI for all HTTP requests, usually the discovery URI.
-      This parameter is no longer used as rootUrl and servicePath are included
-      within the discovery document. (deprecated)
-    future: string, discovery document with future capabilities (deprecated).
-    http: httplib2.Http, An instance of httplib2.Http or something that acts
-      like it that HTTP requests will be made through.
-    developerKey: string, Key for controlling API usage, generated
-      from the API Console.
-    model: Model class instance that serializes and de-serializes requests and
-      responses.
-    requestBuilder: Takes an http request and packages it up to be executed.
-    credentials: oauth2client.Credentials or
-      google.auth.credentials.Credentials, credentials to be used for
-      authentication.
-    client_options: Mapping object or google.api_core.client_options, client
-      options to set user options on the client.
-      (1) The API endpoint should be set through client_options. If API endpoint
-      is not set, `GOOGLE_API_USE_MTLS_ENDPOINT` environment variable can be used
-      to control which endpoint to use.
-      (2) client_cert_source is not supported, client cert should be provided using
-      client_encrypted_cert_source instead. In order to use the provided client
-      cert, `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be
-      set to `true`.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    adc_cert_path: str, client certificate file path to save the application
-      default client certificate for mTLS. This field is required if you want to
-      use the default client certificate. `GOOGLE_API_USE_CLIENT_CERTIFICATE`
-      environment variable must be set to `true` in order to use this field,
-      otherwise this field doesn't nothing.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    adc_key_path: str, client encrypted private key file path to save the
-      application default client encrypted private key for mTLS. This field is
-      required if you want to use the default client certificate.
-      `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be set to
-      `true` in order to use this field, otherwise this field doesn't nothing.
-      More details on the environment variables are here:
-      https://google.aip.dev/auth/4114
-    always_use_jwt_access: Boolean, whether always use self signed JWT for service
-      account credentials. This only applies to
-      google.oauth2.service_account.Credentials.
+    Args:
+      service: string or object, the JSON discovery document describing the API.
+        The value passed in may either be the JSON string or the deserialized
+        JSON.
+      base: string, base URI for all HTTP requests, usually the discovery URI.
+        This parameter is no longer used as rootUrl and servicePath are included
+        within the discovery document. (deprecated)
+      future: string, discovery document with future capabilities (deprecated).
+      http: httplib2.Http, An instance of httplib2.Http or something that acts
+        like it that HTTP requests will be made through.
+      developerKey: string, Key for controlling API usage, generated
+        from the API Console.
+      model: Model class instance that serializes and de-serializes requests and
+        responses.
+      requestBuilder: Takes an http request and packages it up to be executed.
+      credentials: oauth2client.Credentials or
+        google.auth.credentials.Credentials, credentials to be used for
+        authentication.
+      client_options: Mapping object or google.api_core.client_options, client
+        options to set user options on the client.
+        (1) The API endpoint should be set through client_options. If API endpoint
+        is not set, `GOOGLE_API_USE_MTLS_ENDPOINT` environment variable can be used
+        to control which endpoint to use.
+        (2) client_cert_source is not supported, client cert should be provided using
+        client_encrypted_cert_source instead. In order to use the provided client
+        cert, `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be
+        set to `true`.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      adc_cert_path: str, client certificate file path to save the application
+        default client certificate for mTLS. This field is required if you want to
+        use the default client certificate. `GOOGLE_API_USE_CLIENT_CERTIFICATE`
+        environment variable must be set to `true` in order to use this field,
+        otherwise this field doesn't nothing.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      adc_key_path: str, client encrypted private key file path to save the
+        application default client encrypted private key for mTLS. This field is
+        required if you want to use the default client certificate.
+        `GOOGLE_API_USE_CLIENT_CERTIFICATE` environment variable must be set to
+        `true` in order to use this field, otherwise this field doesn't nothing.
+        More details on the environment variables are here:
+        https://google.aip.dev/auth/4114
+      always_use_jwt_access: Boolean, whether always use self signed JWT for service
+        account credentials. This only applies to
+        google.oauth2.service_account.Credentials.
 
-  Returns:
-    A Resource object with methods for interacting with the service.
+    Returns:
+      A Resource object with methods for interacting with the service.
 
-  Raises:
-    google.auth.exceptions.MutualTLSChannelError: if there are any problems
-      setting up mutual TLS channel.
-  """
+    Raises:
+      google.auth.exceptions.MutualTLSChannelError: if there are any problems
+        setting up mutual TLS channel.
+    """
 
     if client_options is None:
         client_options = google.api_core.client_options.ClientOptions()
@@ -522,7 +524,9 @@ def build_from_document(
         ]
         for option, name in banned_options:
             if option is not None:
-                raise ValueError("Arguments http and {} are mutually exclusive".format(name))
+                raise ValueError(
+                    "Arguments http and {} are mutually exclusive".format(name)
+                )
 
     if isinstance(service, str):
         service = json.loads(service)
@@ -562,7 +566,7 @@ def build_from_document(
             if client_options.credentials_file and credentials:
                 raise google.api_core.exceptions.DuplicateCredentialArgs(
                     "client_options.credentials_file and credentials are mutually exclusive."
-            )
+                )
             # Check for credentials file via client options
             if client_options.credentials_file:
                 credentials = _auth.credentials_from_file(
@@ -647,7 +651,9 @@ def build_from_document(
         if "mtlsRootUrl" in service and (
             not client_options or not client_options.api_endpoint
         ):
-            mtls_endpoint = urllib.parse.urljoin(service["mtlsRootUrl"], service["servicePath"])
+            mtls_endpoint = urllib.parse.urljoin(
+                service["mtlsRootUrl"], service["servicePath"]
+            )
             use_mtls_endpoint = os.getenv(GOOGLE_API_USE_MTLS_ENDPOINT, "auto")
 
             if not use_mtls_endpoint in ("never", "auto", "always"):
@@ -681,18 +687,18 @@ def build_from_document(
 def _cast(value, schema_type):
     """Convert value to a string based on JSON Schema type.
 
-  See http://tools.ietf.org/html/draft-zyp-json-schema-03 for more details on
-  JSON Schema.
+    See http://tools.ietf.org/html/draft-zyp-json-schema-03 for more details on
+    JSON Schema.
 
-  Args:
-    value: any, the value to convert
-    schema_type: string, the type that value should be interpreted as
+    Args:
+      value: any, the value to convert
+      schema_type: string, the type that value should be interpreted as
 
-  Returns:
-    A string representation of 'value' based on the schema_type.
-  """
+    Returns:
+      A string representation of 'value' based on the schema_type.
+    """
     if schema_type == "string":
-        if type(value) == type("") or type(value) == type(u""):
+        if type(value) == type("") or type(value) == type(""):
             return value
         else:
             return str(value)
@@ -703,7 +709,7 @@ def _cast(value, schema_type):
     elif schema_type == "boolean":
         return str(bool(value)).lower()
     else:
-        if type(value) == type("") or type(value) == type(u""):
+        if type(value) == type("") or type(value) == type(""):
             return value
         else:
             return str(value)
@@ -712,12 +718,12 @@ def _cast(value, schema_type):
 def _media_size_to_long(maxSize):
     """Convert a string media size, such as 10GB or 3TB into an integer.
 
-  Args:
-    maxSize: string, size as a string, such as 2MB or 7GB.
+    Args:
+      maxSize: string, size as a string, such as 2MB or 7GB.
 
-  Returns:
-    The size as an integer value.
-  """
+    Returns:
+      The size as an integer value.
+    """
     if len(maxSize) < 2:
         return 0
     units = maxSize[-2:].upper()
@@ -731,17 +737,17 @@ def _media_size_to_long(maxSize):
 def _media_path_url_from_info(root_desc, path_url):
     """Creates an absolute media path URL.
 
-  Constructed using the API root URI and service path from the discovery
-  document and the relative path for the API method.
+    Constructed using the API root URI and service path from the discovery
+    document and the relative path for the API method.
 
-  Args:
-    root_desc: Dictionary; the entire original deserialized discovery document.
-    path_url: String; the relative URL for the API method. Relative to the API
-        root, which is specified in the discovery document.
+    Args:
+      root_desc: Dictionary; the entire original deserialized discovery document.
+      path_url: String; the relative URL for the API method. Relative to the API
+          root, which is specified in the discovery document.
 
-  Returns:
-    String; the absolute URI for media upload for the API method.
-  """
+    Returns:
+      String; the absolute URI for media upload for the API method.
+    """
     return "%(root)supload/%(service_path)s%(path)s" % {
         "root": root_desc["rootUrl"],
         "service_path": root_desc["servicePath"],
@@ -752,27 +758,27 @@ def _media_path_url_from_info(root_desc, path_url):
 def _fix_up_parameters(method_desc, root_desc, http_method, schema):
     """Updates parameters of an API method with values specific to this library.
 
-  Specifically, adds whatever global parameters are specified by the API to the
-  parameters for the individual method. Also adds parameters which don't
-  appear in the discovery document, but are available to all discovery based
-  APIs (these are listed in STACK_QUERY_PARAMETERS).
+    Specifically, adds whatever global parameters are specified by the API to the
+    parameters for the individual method. Also adds parameters which don't
+    appear in the discovery document, but are available to all discovery based
+    APIs (these are listed in STACK_QUERY_PARAMETERS).
 
-  SIDE EFFECTS: This updates the parameters dictionary object in the method
-  description.
+    SIDE EFFECTS: This updates the parameters dictionary object in the method
+    description.
 
-  Args:
-    method_desc: Dictionary with metadata describing an API method. Value comes
-        from the dictionary of methods stored in the 'methods' key in the
-        deserialized discovery document.
-    root_desc: Dictionary; the entire original deserialized discovery document.
-    http_method: String; the HTTP method used to call the API method described
-        in method_desc.
-    schema: Object, mapping of schema names to schema descriptions.
+    Args:
+      method_desc: Dictionary with metadata describing an API method. Value comes
+          from the dictionary of methods stored in the 'methods' key in the
+          deserialized discovery document.
+      root_desc: Dictionary; the entire original deserialized discovery document.
+      http_method: String; the HTTP method used to call the API method described
+          in method_desc.
+      schema: Object, mapping of schema names to schema descriptions.
 
-  Returns:
-    The updated Dictionary stored in the 'parameters' key of the method
-        description dictionary.
-  """
+    Returns:
+      The updated Dictionary stored in the 'parameters' key of the method
+          description dictionary.
+    """
     parameters = method_desc.setdefault("parameters", {})
 
     # Add in the parameters common to all methods.
@@ -796,31 +802,31 @@ def _fix_up_parameters(method_desc, root_desc, http_method, schema):
 def _fix_up_media_upload(method_desc, root_desc, path_url, parameters):
     """Adds 'media_body' and 'media_mime_type' parameters if supported by method.
 
-  SIDE EFFECTS: If there is a 'mediaUpload' in the method description, adds
-  'media_upload' key to parameters.
+    SIDE EFFECTS: If there is a 'mediaUpload' in the method description, adds
+    'media_upload' key to parameters.
 
-  Args:
-    method_desc: Dictionary with metadata describing an API method. Value comes
-        from the dictionary of methods stored in the 'methods' key in the
-        deserialized discovery document.
-    root_desc: Dictionary; the entire original deserialized discovery document.
-    path_url: String; the relative URL for the API method. Relative to the API
-        root, which is specified in the discovery document.
-    parameters: A dictionary describing method parameters for method described
-        in method_desc.
+    Args:
+      method_desc: Dictionary with metadata describing an API method. Value comes
+          from the dictionary of methods stored in the 'methods' key in the
+          deserialized discovery document.
+      root_desc: Dictionary; the entire original deserialized discovery document.
+      path_url: String; the relative URL for the API method. Relative to the API
+          root, which is specified in the discovery document.
+      parameters: A dictionary describing method parameters for method described
+          in method_desc.
 
-  Returns:
-    Triple (accept, max_size, media_path_url) where:
-      - accept is a list of strings representing what content types are
-        accepted for media upload. Defaults to empty list if not in the
-        discovery document.
-      - max_size is a long representing the max size in bytes allowed for a
-        media upload. Defaults to 0L if not in the discovery document.
-      - media_path_url is a String; the absolute URI for media upload for the
-        API method. Constructed using the API root URI and service path from
-        the discovery document and the relative path for the API method. If
-        media upload is not supported, this is None.
-  """
+    Returns:
+      Triple (accept, max_size, media_path_url) where:
+        - accept is a list of strings representing what content types are
+          accepted for media upload. Defaults to empty list if not in the
+          discovery document.
+        - max_size is a long representing the max size in bytes allowed for a
+          media upload. Defaults to 0L if not in the discovery document.
+        - media_path_url is a String; the absolute URI for media upload for the
+          API method. Constructed using the API root URI and service path from
+          the discovery document and the relative path for the API method. If
+          media upload is not supported, this is None.
+    """
     media_upload = method_desc.get("mediaUpload", {})
     accept = media_upload.get("accept", [])
     max_size = _media_size_to_long(media_upload.get("maxSize", ""))
@@ -837,35 +843,35 @@ def _fix_up_media_upload(method_desc, root_desc, path_url, parameters):
 def _fix_up_method_description(method_desc, root_desc, schema):
     """Updates a method description in a discovery document.
 
-  SIDE EFFECTS: Changes the parameters dictionary in the method description with
-  extra parameters which are used locally.
+    SIDE EFFECTS: Changes the parameters dictionary in the method description with
+    extra parameters which are used locally.
 
-  Args:
-    method_desc: Dictionary with metadata describing an API method. Value comes
-        from the dictionary of methods stored in the 'methods' key in the
-        deserialized discovery document.
-    root_desc: Dictionary; the entire original deserialized discovery document.
-    schema: Object, mapping of schema names to schema descriptions.
+    Args:
+      method_desc: Dictionary with metadata describing an API method. Value comes
+          from the dictionary of methods stored in the 'methods' key in the
+          deserialized discovery document.
+      root_desc: Dictionary; the entire original deserialized discovery document.
+      schema: Object, mapping of schema names to schema descriptions.
 
-  Returns:
-    Tuple (path_url, http_method, method_id, accept, max_size, media_path_url)
-    where:
-      - path_url is a String; the relative URL for the API method. Relative to
-        the API root, which is specified in the discovery document.
-      - http_method is a String; the HTTP method used to call the API method
-        described in the method description.
-      - method_id is a String; the name of the RPC method associated with the
-        API method, and is in the method description in the 'id' key.
-      - accept is a list of strings representing what content types are
-        accepted for media upload. Defaults to empty list if not in the
-        discovery document.
-      - max_size is a long representing the max size in bytes allowed for a
-        media upload. Defaults to 0L if not in the discovery document.
-      - media_path_url is a String; the absolute URI for media upload for the
-        API method. Constructed using the API root URI and service path from
-        the discovery document and the relative path for the API method. If
-        media upload is not supported, this is None.
-  """
+    Returns:
+      Tuple (path_url, http_method, method_id, accept, max_size, media_path_url)
+      where:
+        - path_url is a String; the relative URL for the API method. Relative to
+          the API root, which is specified in the discovery document.
+        - http_method is a String; the HTTP method used to call the API method
+          described in the method description.
+        - method_id is a String; the name of the RPC method associated with the
+          API method, and is in the method description in the 'id' key.
+        - accept is a list of strings representing what content types are
+          accepted for media upload. Defaults to empty list if not in the
+          discovery document.
+        - max_size is a long representing the max size in bytes allowed for a
+          media upload. Defaults to 0L if not in the discovery document.
+        - media_path_url is a String; the absolute URI for media upload for the
+          API method. Constructed using the API root URI and service path from
+          the discovery document and the relative path for the API method. If
+          media upload is not supported, this is None.
+    """
     path_url = method_desc["path"]
     http_method = method_desc["httpMethod"]
     method_id = method_desc["id"]
@@ -902,38 +908,38 @@ def _urljoin(base, url):
 class ResourceMethodParameters(object):
     """Represents the parameters associated with a method.
 
-  Attributes:
-    argmap: Map from method parameter name (string) to query parameter name
-        (string).
-    required_params: List of required parameters (represented by parameter
-        name as string).
-    repeated_params: List of repeated parameters (represented by parameter
-        name as string).
-    pattern_params: Map from method parameter name (string) to regular
-        expression (as a string). If the pattern is set for a parameter, the
-        value for that parameter must match the regular expression.
-    query_params: List of parameters (represented by parameter name as string)
-        that will be used in the query string.
-    path_params: Set of parameters (represented by parameter name as string)
-        that will be used in the base URL path.
-    param_types: Map from method parameter name (string) to parameter type. Type
-        can be any valid JSON schema type; valid values are 'any', 'array',
-        'boolean', 'integer', 'number', 'object', or 'string'. Reference:
-        http://tools.ietf.org/html/draft-zyp-json-schema-03#section-5.1
-    enum_params: Map from method parameter name (string) to list of strings,
-       where each list of strings is the list of acceptable enum values.
-  """
+    Attributes:
+      argmap: Map from method parameter name (string) to query parameter name
+          (string).
+      required_params: List of required parameters (represented by parameter
+          name as string).
+      repeated_params: List of repeated parameters (represented by parameter
+          name as string).
+      pattern_params: Map from method parameter name (string) to regular
+          expression (as a string). If the pattern is set for a parameter, the
+          value for that parameter must match the regular expression.
+      query_params: List of parameters (represented by parameter name as string)
+          that will be used in the query string.
+      path_params: Set of parameters (represented by parameter name as string)
+          that will be used in the base URL path.
+      param_types: Map from method parameter name (string) to parameter type. Type
+          can be any valid JSON schema type; valid values are 'any', 'array',
+          'boolean', 'integer', 'number', 'object', or 'string'. Reference:
+          http://tools.ietf.org/html/draft-zyp-json-schema-03#section-5.1
+      enum_params: Map from method parameter name (string) to list of strings,
+         where each list of strings is the list of acceptable enum values.
+    """
 
     def __init__(self, method_desc):
         """Constructor for ResourceMethodParameters.
 
-    Sets default values and defers to set_parameters to populate.
+        Sets default values and defers to set_parameters to populate.
 
-    Args:
-      method_desc: Dictionary with metadata describing an API method. Value
-          comes from the dictionary of methods stored in the 'methods' key in
-          the deserialized discovery document.
-    """
+        Args:
+          method_desc: Dictionary with metadata describing an API method. Value
+              comes from the dictionary of methods stored in the 'methods' key in
+              the deserialized discovery document.
+        """
         self.argmap = {}
         self.required_params = []
         self.repeated_params = []
@@ -950,14 +956,14 @@ class ResourceMethodParameters(object):
     def set_parameters(self, method_desc):
         """Populates maps and lists based on method description.
 
-    Iterates through each parameter for the method and parses the values from
-    the parameter dictionary.
+        Iterates through each parameter for the method and parses the values from
+        the parameter dictionary.
 
-    Args:
-      method_desc: Dictionary with metadata describing an API method. Value
-          comes from the dictionary of methods stored in the 'methods' key in
-          the deserialized discovery document.
-    """
+        Args:
+          method_desc: Dictionary with metadata describing an API method. Value
+              comes from the dictionary of methods stored in the 'methods' key in
+              the deserialized discovery document.
+        """
         parameters = method_desc.get("parameters", {})
         sorted_parameters = OrderedDict(sorted(parameters.items()))
         for arg, desc in sorted_parameters.items():
@@ -992,13 +998,13 @@ class ResourceMethodParameters(object):
 def createMethod(methodName, methodDesc, rootDesc, schema):
     """Creates a method for attaching to a Resource.
 
-  Args:
-    methodName: string, name of the method to use.
-    methodDesc: object, fragment of deserialized discovery document that
-      describes the method.
-    rootDesc: object, the entire deserialized discovery document.
-    schema: object, mapping of schema names to schema descriptions.
-  """
+    Args:
+      methodName: string, name of the method to use.
+      methodDesc: object, fragment of deserialized discovery document that
+        describes the method.
+      rootDesc: object, the entire deserialized discovery document.
+      schema: object, mapping of schema names to schema descriptions.
+    """
     methodName = fix_method_name(methodName)
     (
         pathUrl,
@@ -1016,7 +1022,7 @@ def createMethod(methodName, methodDesc, rootDesc, schema):
 
         for name in kwargs:
             if name not in parameters.argmap:
-                raise TypeError('Got an unexpected keyword argument {}'.format(name))
+                raise TypeError("Got an unexpected keyword argument {}".format(name))
 
         # Remove args that have a value of None.
         keys = list(kwargs.keys())
@@ -1256,28 +1262,28 @@ def createNextMethod(
 ):
     """Creates any _next methods for attaching to a Resource.
 
-  The _next methods allow for easy iteration through list() responses.
+    The _next methods allow for easy iteration through list() responses.
 
-  Args:
-    methodName: string, name of the method to use.
-    pageTokenName: string, name of request page token field.
-    nextPageTokenName: string, name of response page token field.
-    isPageTokenParameter: Boolean, True if request page token is a query
-        parameter, False if request page token is a field of the request body.
-  """
+    Args:
+      methodName: string, name of the method to use.
+      pageTokenName: string, name of request page token field.
+      nextPageTokenName: string, name of response page token field.
+      isPageTokenParameter: Boolean, True if request page token is a query
+          parameter, False if request page token is a field of the request body.
+    """
     methodName = fix_method_name(methodName)
 
     def methodNext(self, previous_request, previous_response):
         """Retrieves the next page of results.
 
-Args:
-  previous_request: The request for the previous page. (required)
-  previous_response: The response from the request for the previous page. (required)
+        Args:
+          previous_request: The request for the previous page. (required)
+          previous_response: The response from the request for the previous page. (required)
 
-Returns:
-  A request object that you can call 'execute()' on to request the next
-  page. Returns None if there are no more items in the collection.
-    """
+        Returns:
+          A request object that you can call 'execute()' on to request the next
+          page. Returns None if there are no more items in the collection.
+        """
         # Retrieve nextPageToken from previous_response
         # Use as pageToken in previous_request to create new request.
 
@@ -1301,7 +1307,7 @@ Returns:
             request.body = model.serialize(body)
             request.body_size = len(request.body)
             if "content-length" in request.headers:
-              del request.headers["content-length"]
+                del request.headers["content-length"]
             logger.debug("Next page request body: %s %s" % (methodName, body))
 
         return request
@@ -1325,21 +1331,21 @@ class Resource(object):
     ):
         """Build a Resource from the API description.
 
-    Args:
-      http: httplib2.Http, Object to make http requests with.
-      baseUrl: string, base URL for the API. All requests are relative to this
-          URI.
-      model: googleapiclient.Model, converts to and from the wire format.
-      requestBuilder: class or callable that instantiates an
-          googleapiclient.HttpRequest object.
-      developerKey: string, key obtained from
-          https://code.google.com/apis/console
-      resourceDesc: object, section of deserialized discovery document that
-          describes a resource. Note that the top level discovery document
-          is considered a resource.
-      rootDesc: object, the entire deserialized discovery document.
-      schema: object, mapping of schema names to schema descriptions.
-    """
+        Args:
+          http: httplib2.Http, Object to make http requests with.
+          baseUrl: string, base URL for the API. All requests are relative to this
+              URI.
+          model: googleapiclient.Model, converts to and from the wire format.
+          requestBuilder: class or callable that instantiates an
+              googleapiclient.HttpRequest object.
+          developerKey: string, key obtained from
+              https://code.google.com/apis/console
+          resourceDesc: object, section of deserialized discovery document that
+              describes a resource. Note that the top level discovery document
+              is considered a resource.
+          rootDesc: object, the entire deserialized discovery document.
+          schema: object, mapping of schema names to schema descriptions.
+        """
         self._dynamic_attrs = []
 
         self._http = http
@@ -1356,19 +1362,19 @@ class Resource(object):
     def _set_dynamic_attr(self, attr_name, value):
         """Sets an instance attribute and tracks it in a list of dynamic attributes.
 
-    Args:
-      attr_name: string; The name of the attribute to be set
-      value: The value being set on the object and tracked in the dynamic cache.
-    """
+        Args:
+          attr_name: string; The name of the attribute to be set
+          value: The value being set on the object and tracked in the dynamic cache.
+        """
         self._dynamic_attrs.append(attr_name)
         self.__dict__[attr_name] = value
 
     def __getstate__(self):
         """Trim the state down to something that can be pickled.
 
-    Uses the fact that the instance variable _dynamic_attrs holds attrs that
-    will be wiped and restored on pickle serialization.
-    """
+        Uses the fact that the instance variable _dynamic_attrs holds attrs that
+        will be wiped and restored on pickle serialization.
+        """
         state_dict = copy.copy(self.__dict__)
         for dynamic_attr in self._dynamic_attrs:
             del state_dict[dynamic_attr]
@@ -1378,13 +1384,12 @@ class Resource(object):
     def __setstate__(self, state):
         """Reconstitute the state of the object from being pickled.
 
-    Uses the fact that the instance variable _dynamic_attrs holds attrs that
-    will be wiped and restored on pickle serialization.
-    """
+        Uses the fact that the instance variable _dynamic_attrs holds attrs that
+        will be wiped and restored on pickle serialization.
+        """
         self.__dict__.update(state)
         self._dynamic_attrs = []
         self._set_service_methods()
-
 
     def __enter__(self):
         return self
@@ -1415,17 +1420,17 @@ class Resource(object):
             def new_batch_http_request(callback=None):
                 """Create a BatchHttpRequest object based on the discovery document.
 
-        Args:
-          callback: callable, A callback to be called for each response, of the
-            form callback(id, response, exception). The first parameter is the
-            request id, and the second is the deserialized response object. The
-            third is an apiclient.errors.HttpError exception object if an HTTP
-            error occurred while processing the request, or None if no error
-            occurred.
+                Args:
+                  callback: callable, A callback to be called for each response, of the
+                    form callback(id, response, exception). The first parameter is the
+                    request id, and the second is the deserialized response object. The
+                    third is an apiclient.errors.HttpError exception object if an HTTP
+                    error occurred while processing the request, or None if no error
+                    occurred.
 
-        Returns:
-          A BatchHttpRequest object based on the discovery document.
-        """
+                Returns:
+                  A BatchHttpRequest object based on the discovery document.
+                """
                 return BatchHttpRequest(callback=callback, batch_uri=batch_uri)
 
             self._set_dynamic_attr("new_batch_http_request", new_batch_http_request)
@@ -1456,11 +1461,11 @@ class Resource(object):
             def createResourceMethod(methodName, methodDesc):
                 """Create a method on the Resource to access a nested Resource.
 
-        Args:
-          methodName: string, name of the method to use.
-          methodDesc: object, fragment of deserialized discovery document that
-            describes the method.
-        """
+                Args:
+                  methodName: string, name of the method to use.
+                  methodDesc: object, fragment of deserialized discovery document that
+                    describes the method.
+                """
                 methodName = fix_method_name(methodName)
 
                 def methodResource(self):
@@ -1521,13 +1526,13 @@ class Resource(object):
 def _findPageTokenName(fields):
     """Search field names for one like a page token.
 
-  Args:
-    fields: container of string, names of fields.
+    Args:
+      fields: container of string, names of fields.
 
-  Returns:
-    First name that is either 'pageToken' or 'nextPageToken' if one exists,
-    otherwise None.
-  """
+    Returns:
+      First name that is either 'pageToken' or 'nextPageToken' if one exists,
+      otherwise None.
+    """
     return next(
         (tokenName for tokenName in _PAGE_TOKEN_NAMES if tokenName in fields), None
     )
@@ -1536,17 +1541,17 @@ def _findPageTokenName(fields):
 def _methodProperties(methodDesc, schema, name):
     """Get properties of a field in a method description.
 
-  Args:
-    methodDesc: object, fragment of deserialized discovery document that
-      describes the method.
-    schema: object, mapping of schema names to schema descriptions.
-    name: string, name of top-level field in method description.
+    Args:
+      methodDesc: object, fragment of deserialized discovery document that
+        describes the method.
+      schema: object, mapping of schema names to schema descriptions.
+      name: string, name of top-level field in method description.
 
-  Returns:
-    Object representing fragment of deserialized discovery document
-    corresponding to 'properties' field of object corresponding to named field
-    in method description, if it exists, otherwise empty dict.
-  """
+    Returns:
+      Object representing fragment of deserialized discovery document
+      corresponding to 'properties' field of object corresponding to named field
+      in method description, if it exists, otherwise empty dict.
+    """
     desc = methodDesc.get(name, {})
     if "$ref" in desc:
         desc = schema.get(desc["$ref"], {})
