@@ -15,7 +15,7 @@
 """File based cache for the discovery document.
 
 The cache is stored in a single file so that multiple processes can
-share the same cache. It locks the file whenever accesing to the
+share the same cache. It locks the file whenever accessing to the
 file. When the cache content is corrupted, it will be initialized with
 an empty cache.
 """
@@ -27,7 +27,6 @@ import json
 import logging
 import os
 import tempfile
-import threading
 
 try:
     from oauth2client.contrib.locked_file import LockedFile
@@ -58,8 +57,8 @@ def _to_timestamp(date):
         # See also: https://docs.python.org/2/library/datetime.html
         delta = date - EPOCH
         return (
-            delta.microseconds + (delta.seconds + delta.days * 24 * 3600) * 10 ** 6
-        ) / 10 ** 6
+            delta.microseconds + (delta.seconds + delta.days * 24 * 3600) * 10**6
+        ) / 10**6
 
 
 def _read_or_initialize_cache(f):
@@ -82,9 +81,9 @@ class Cache(base.Cache):
     def __init__(self, max_age):
         """Constructor.
 
-      Args:
-        max_age: Cache expiration in seconds.
-      """
+        Args:
+          max_age: Cache expiration in seconds.
+        """
         self._max_age = max_age
         self._file = os.path.join(tempfile.gettempdir(), FILENAME)
         f = LockedFile(self._file, "a+", "r")
