@@ -13,8 +13,23 @@
 # limitations under the License.
 
 import os
-import nox
 import shutil
+
+import nox
+
+BLACK_VERSION = "black==22.3.0"
+ISORT_VERSION = "isort==5.10.1"
+BLACK_PATHS = [
+    "apiclient",
+    "googleapiclient",
+    "scripts",
+    "tests",
+    "describe.py",
+    "expandsymlinks.py",
+    "noxfile.py",
+    "owlbot.py",
+    "setup.py",
+]
 
 test_dependencies = [
     "django>=2.0.0",
@@ -27,7 +42,6 @@ test_dependencies = [
     "pytest-cov",
     "webtest",
     "coverage",
-    "mock",
 ]
 
 
@@ -45,7 +59,27 @@ def lint(session):
     )
 
 
-@nox.session(python=["3.6", "3.7", "3.8", "3.9", "3.10"])
+@nox.session(python="3.8")
+def format(session):
+    """
+    Run isort to sort imports. Then run black
+    to format code to uniform standard.
+    """
+    session.install(BLACK_VERSION, ISORT_VERSION)
+    # Use the --fss option to sort imports using strict alphabetical order.
+    # See https://pycqa.github.io/isort/docs/configuration/options.html#force-sort-within-sections
+    session.run(
+        "isort",
+        "--fss",
+        *BLACK_PATHS,
+    )
+    session.run(
+        "black",
+        *BLACK_PATHS,
+    )
+
+
+@nox.session(python=["3.7", "3.8", "3.9", "3.10"])
 @nox.parametrize(
     "oauth2client",
     [
