@@ -20,7 +20,7 @@ Tags: adclients.list
 """
 from __future__ import print_function
 
-__author__ = 'sgomes@google.com (Sérgio Gomes)'
+__author__ = "sgomes@google.com (Sérgio Gomes)"
 
 import sys
 
@@ -31,30 +31,47 @@ MAX_PAGE_SIZE = 50
 
 
 def main(argv):
-  # Authenticate and construct service.
-  service, flags = sample_tools.init(
-      argv, 'adexchangeseller', 'v1.1', __doc__, __file__, parents=[],
-      scope='https://www.googleapis.com/auth/adexchange.seller.readonly')
+    # Authenticate and construct service.
+    service, flags = sample_tools.init(
+        argv,
+        "adexchangeseller",
+        "v1.1",
+        __doc__,
+        __file__,
+        parents=[],
+        scope="https://www.googleapis.com/auth/adexchange.seller.readonly",
+    )
 
-  try:
-    # Retrieve ad client list in pages and display data as we receive it.
-    request = service.adclients().list(maxResults=MAX_PAGE_SIZE)
+    try:
+        # Retrieve ad client list in pages and display data as we receive it.
+        request = service.adclients().list(maxResults=MAX_PAGE_SIZE)
 
-    while request is not None:
-      result = request.execute()
-      ad_clients = result['items']
-      for ad_client in ad_clients:
-        print(('Ad client for product "%s" with ID "%s" was found. '
-               % (ad_client['productCode'], ad_client['id'])))
+        while request is not None:
+            result = request.execute()
+            ad_clients = result["items"]
+            for ad_client in ad_clients:
+                print(
+                    (
+                        'Ad client for product "%s" with ID "%s" was found. '
+                        % (ad_client["productCode"], ad_client["id"])
+                    )
+                )
 
-        print(('\tSupports reporting: %s' %
-               (ad_client['supportsReporting'] and 'Yes' or 'No')))
+                print(
+                    (
+                        "\tSupports reporting: %s"
+                        % (ad_client["supportsReporting"] and "Yes" or "No")
+                    )
+                )
 
-      request = service.adclients().list_next(request, result)
+            request = service.adclients().list_next(request, result)
 
-  except client.AccessTokenRefreshError:
-    print ('The credentials have been revoked or expired, please re-run the '
-           'application to re-authorize')
+    except client.AccessTokenRefreshError:
+        print(
+            "The credentials have been revoked or expired, please re-run the "
+            "application to re-authorize"
+        )
 
-if __name__ == '__main__':
-  main(sys.argv)
+
+if __name__ == "__main__":
+    main(sys.argv)
