@@ -20,9 +20,8 @@ Tags: alerts.list
 """
 from __future__ import print_function
 
-__author__ = 'sgomes@google.com (Sérgio Gomes)'
+__author__ = "sgomes@google.com (Sérgio Gomes)"
 
-import argparse
 import sys
 
 from googleapiclient import sample_tools
@@ -30,27 +29,40 @@ from oauth2client import client
 
 
 def main(argv):
-  # Authenticate and construct service.
-  service, flags = sample_tools.init(
-      argv, 'adexchangeseller', 'v1.1', __doc__, __file__, parents=[],
-      scope='https://www.googleapis.com/auth/adexchange.seller.readonly')
+    # Authenticate and construct service.
+    service, flags = sample_tools.init(
+        argv,
+        "adexchangeseller",
+        "v1.1",
+        __doc__,
+        __file__,
+        parents=[],
+        scope="https://www.googleapis.com/auth/adexchange.seller.readonly",
+    )
 
-  try:
-    # Retrieve alerts list in pages and display data as we receive it.
-    request = service.alerts().list()
+    try:
+        # Retrieve alerts list in pages and display data as we receive it.
+        request = service.alerts().list()
 
-    if request is not None:
-      result = request.execute()
-      if 'items' in result:
-        alerts = result['items']
-        for alert in alerts:
-          print(('Alert id "%s" with severity "%s" and type "%s" was found. '
-                 % (alert['id'], alert['severity'], alert['type'])))
-      else:
-        print('No alerts found!')
-  except client.AccessTokenRefreshError:
-    print ('The credentials have been revoked or expired, please re-run the '
-           'application to re-authorize')
+        if request is not None:
+            result = request.execute()
+            if "items" in result:
+                alerts = result["items"]
+                for alert in alerts:
+                    print(
+                        (
+                            'Alert id "%s" with severity "%s" and type "%s" was found. '
+                            % (alert["id"], alert["severity"], alert["type"])
+                        )
+                    )
+            else:
+                print("No alerts found!")
+    except client.AccessTokenRefreshError:
+        print(
+            "The credentials have been revoked or expired, please re-run the "
+            "application to re-authorize"
+        )
 
-if __name__ == '__main__':
-  main(sys.argv)
+
+if __name__ == "__main__":
+    main(sys.argv)
