@@ -20,9 +20,8 @@ Tags: preferreddeals.list
 """
 from __future__ import print_function
 
-__author__ = 'sgomes@google.com (Sérgio Gomes)'
+__author__ = "sgomes@google.com (Sérgio Gomes)"
 
-import argparse
 import sys
 
 from googleapiclient import sample_tools
@@ -30,35 +29,44 @@ from oauth2client import client
 
 
 def main(argv):
-  # Authenticate and construct service.
-  service, flags = sample_tools.init(
-      argv, 'adexchangeseller', 'v1.1', __doc__, __file__, parents=[],
-      scope='https://www.googleapis.com/auth/adexchange.seller.readonly')
+    # Authenticate and construct service.
+    service, flags = sample_tools.init(
+        argv,
+        "adexchangeseller",
+        "v1.1",
+        __doc__,
+        __file__,
+        parents=[],
+        scope="https://www.googleapis.com/auth/adexchange.seller.readonly",
+    )
 
-  try:
-    # Retrieve preferred deals list in pages and display data as we receive it.
-    request = service.preferreddeals().list()
+    try:
+        # Retrieve preferred deals list in pages and display data as we receive it.
+        request = service.preferreddeals().list()
 
-    if request is not None:
-      result = request.execute()
-      if 'items' in result:
-        deals = result['items']
-        for deal in deals:
-          output = 'Deal id "%s" ' % deal['id']
+        if request is not None:
+            result = request.execute()
+            if "items" in result:
+                deals = result["items"]
+                for deal in deals:
+                    output = 'Deal id "%s" ' % deal["id"]
 
-          if 'advertiserName' in deal:
-            output += 'for advertiser "%s" ' % deal['advertiserName']
+                    if "advertiserName" in deal:
+                        output += 'for advertiser "%s" ' % deal["advertiserName"]
 
-          if 'buyerNetworkName' in deal:
-            output += 'on network "%s" ' % deal['buyerNetworkName']
+                    if "buyerNetworkName" in deal:
+                        output += 'on network "%s" ' % deal["buyerNetworkName"]
 
-          output += 'was found.'
-          print(output)
-      else:
-        print('No preferred deals found!')
-  except client.AccessTokenRefreshError:
-    print ('The credentials have been revoked or expired, please re-run the '
-           'application to re-authorize')
+                    output += "was found."
+                    print(output)
+            else:
+                print("No preferred deals found!")
+    except client.AccessTokenRefreshError:
+        print(
+            "The credentials have been revoked or expired, please re-run the "
+            "application to re-authorize"
+        )
 
-if __name__ == '__main__':
-  main(sys.argv)
+
+if __name__ == "__main__":
+    main(sys.argv)
