@@ -22,13 +22,11 @@ be obtained from the Developer Console <https://code.google.com/apis/console/>
 and save them as 'client_secrets.json' in the project directory.
 """
 
-__author__ = 'jcgregorio@google.com (Joe Gregorio)'
+__author__ = "jcgregorio@google.com (Joe Gregorio)"
 
 
 import httplib2
-import logging
 import os
-import pickle
 
 from googleapiclient import discovery
 from oauth2client import client
@@ -42,17 +40,19 @@ import jinja2
 JINJA_ENVIRONMENT = jinja2.Environment(
     loader=jinja2.FileSystemLoader(os.path.dirname(__file__)),
     autoescape=True,
-    extensions=['jinja2.ext.autoescape'])
+    extensions=["jinja2.ext.autoescape"],
+)
 
 # CLIENT_SECRETS, name of a file containing the OAuth 2.0 information for this
 # application, including client_id and client_secret, which are found
 # on the API Access tab on the Google APIs
 # Console <http://code.google.com/apis/console>
-CLIENT_SECRETS = os.path.join(os.path.dirname(__file__), 'client_secrets.json')
+CLIENT_SECRETS = os.path.join(os.path.dirname(__file__), "client_secrets.json")
 
 # Helpful message to display in the browser if the CLIENT_SECRETS file
 # is missing.
-MISSING_CLIENT_SECRETS_MESSAGE = """
+MISSING_CLIENT_SECRETS_MESSAGE = (
+    """
 <h1>Warning: Please configure OAuth 2.0</h1>
 <p>
 To make this sample run you will need to populate the client_secrets.json file
@@ -64,47 +64,50 @@ found at:
 <p>with information found on the <a
 href="https://code.google.com/apis/console">APIs Console</a>.
 </p>
-""" % CLIENT_SECRETS
+"""
+    % CLIENT_SECRETS
+)
 
 
 http = httplib2.Http(memcache)
 service = discovery.build("plus", "v1", http=http)
 decorator = appengine.oauth2decorator_from_clientsecrets(
     CLIENT_SECRETS,
-    scope='https://www.googleapis.com/auth/plus.me',
-    message=MISSING_CLIENT_SECRETS_MESSAGE)
+    scope="https://www.googleapis.com/auth/plus.me",
+    message=MISSING_CLIENT_SECRETS_MESSAGE,
+)
+
 
 class MainHandler(webapp2.RequestHandler):
-
-  @decorator.oauth_aware
-  def get(self):
-    variables = {
-        'url': decorator.authorize_url(),
-        'has_credentials': decorator.has_credentials()
+    @decorator.oauth_aware
+    def get(self):
+        variables = {
+            "url": decorator.authorize_url(),
+            "has_credentials": decorator.has_credentials(),
         }
-    template = JINJA_ENVIRONMENT.get_template('grant.html')
-    self.response.write(template.render(variables))
+        template = JINJA_ENVIRONMENT.get_template("grant.html")
+        self.response.write(template.render(variables))
 
 
 class AboutHandler(webapp2.RequestHandler):
+    @decorator.oauth_required
+    def get(self):
+        try:
+            http = decorator.http()
+            user = service.people().get(userId="me").execute(http=http)
+            text = "Hello, %s!" % user["displayName"]
 
-  @decorator.oauth_required
-  def get(self):
-    try:
-      http = decorator.http()
-      user = service.people().get(userId='me').execute(http=http)
-      text = 'Hello, %s!' % user['displayName']
-
-      template = JINJA_ENVIRONMENT.get_template('welcome.html')
-      self.response.write(template.render({'text': text }))
-    except client.AccessTokenRefreshError:
-      self.redirect('/')
+            template = JINJA_ENVIRONMENT.get_template("welcome.html")
+            self.response.write(template.render({"text": text}))
+        except client.AccessTokenRefreshError:
+            self.redirect("/")
 
 
 app = webapp2.WSGIApplication(
     [
-     ('/', MainHandler),
-     ('/about', AboutHandler),
-     (decorator.callback_path, decorator.callback_handler()),
+        ("/", MainHandler),
+        ("/about", AboutHandler),
+        (decorator.callback_path, decorator.callback_handler()),
     ],
-    debug=True)
+    debug=True,
+)
