@@ -35,7 +35,7 @@ To get detailed log output run:
 """
 from __future__ import print_function
 
-__author__ = 'rahulpaul@google.com (Rahul Paul)'
+__author__ = "rahulpaul@google.com (Rahul Paul)"
 
 import pprint
 import re
@@ -46,37 +46,50 @@ from googleapiclient import sample_tools
 
 
 def main(argv):
-  # Authenticate and construct service.
-  service, flags = sample_tools.init(
-      argv, 'audit', 'v1', __doc__, __file__,
-      scope='https://www.googleapis.com/auth/apps/reporting/audit.readonly')
+    # Authenticate and construct service.
+    service, flags = sample_tools.init(
+        argv,
+        "audit",
+        "v1",
+        __doc__,
+        __file__,
+        scope="https://www.googleapis.com/auth/apps/reporting/audit.readonly",
+    )
 
-  try:
-    activities = service.activities()
+    try:
+        activities = service.activities()
 
-    # Retrieve the first two activities
-    print('Retrieving the first 2 activities...')
-    activity_list = activities.list(
-        applicationId='207535951991', customerId='C01rv1wm7', maxResults='2',
-        actorEmail='admin@enterprise-audit-clientlib.com').execute()
-    pprint.pprint(activity_list)
+        # Retrieve the first two activities
+        print("Retrieving the first 2 activities...")
+        activity_list = activities.list(
+            applicationId="207535951991",
+            customerId="C01rv1wm7",
+            maxResults="2",
+            actorEmail="admin@enterprise-audit-clientlib.com",
+        ).execute()
+        pprint.pprint(activity_list)
 
-    # Now retrieve the next 2 events
-    match = re.search('(?<=continuationToken=).+$', activity_list['next'])
-    if match is not None:
-      next_token = match.group(0)
+        # Now retrieve the next 2 events
+        match = re.search("(?<=continuationToken=).+$", activity_list["next"])
+        if match is not None:
+            next_token = match.group(0)
 
-      print('\nRetrieving the next 2 activities...')
-      activity_list = activities.list(
-          applicationId='207535951991', customerId='C01rv1wm7',
-          maxResults='2', actorEmail='admin@enterprise-audit-clientlib.com',
-          continuationToken=next_token).execute()
-      pprint.pprint(activity_list)
+            print("\nRetrieving the next 2 activities...")
+            activity_list = activities.list(
+                applicationId="207535951991",
+                customerId="C01rv1wm7",
+                maxResults="2",
+                actorEmail="admin@enterprise-audit-clientlib.com",
+                continuationToken=next_token,
+            ).execute()
+            pprint.pprint(activity_list)
 
-  except client.AccessTokenRefreshError:
-    print ('The credentials have been revoked or expired, please re-run'
-      'the application to re-authorize')
+    except client.AccessTokenRefreshError:
+        print(
+            "The credentials have been revoked or expired, please re-run"
+            "the application to re-authorize"
+        )
 
-if __name__ == '__main__':
-  main(sys.argv)
 
+if __name__ == "__main__":
+    main(sys.argv)
