@@ -20,7 +20,7 @@
 Command-line application that does a search.
 """
 
-__author__ = 'jcgregorio@google.com (Joe Gregorio)'
+__author__ = "jcgregorio@google.com (Joe Gregorio)"
 
 import pprint
 
@@ -28,17 +28,23 @@ from googleapiclient.discovery import build
 
 
 def main():
-  # Build a service object for interacting with the API. Visit
-  # the Google APIs Console <http://code.google.com/apis/console>
-  # to get an API key for your own application.
-  service = build("customsearch", "v1",
-            developerKey="AIzaSyDRRpR3GS1F1_jKNNM9HCNd2wJQyPG3oN0")
+    # Build a service object for interacting with the API. Visit
+    # the Google APIs Console <http://code.google.com/apis/console>
+    # to get an API key for your own application.
+    service = build(
+        "customsearch", "v1", developerKey="<YOUR DEVELOPER KEY>"
+    )
 
-  res = service.cse().list(
-      q='lectures',
-      cx='017576662512468239146:omuauf_lfve',
-    ).execute()
-  pprint.pprint(res)
+    res = (
+        service.cse()
+        .list(
+            q="lectures",
+            cx="017576662512468239146:omuauf_lfve",
+        )
+        .execute()
+    )
+    pprint.pprint(res)
 
-if __name__ == '__main__':
-  main()
+
+if __name__ == "__main__":
+    main()
