@@ -20,38 +20,43 @@ from oauth2client.contrib.appengine import OAuth2Decorator
 
 import settings
 
-decorator = OAuth2Decorator(client_id=settings.CLIENT_ID,
-                            client_secret=settings.CLIENT_SECRET,
-                            scope=settings.SCOPE)
-service = build('tasks', 'v1')
+decorator = OAuth2Decorator(
+    client_id=settings.CLIENT_ID,
+    client_secret=settings.CLIENT_SECRET,
+    scope=settings.SCOPE,
+)
+service = build("tasks", "v1")
 
 
 class MainHandler(webapp2.RequestHandler):
+    def render_response(self, template, **context):
+        renderer = jinja2.get_jinja2(app=self.app)
+        rendered_value = renderer.render_template(template, **context)
+        self.response.write(rendered_value)
 
-  def render_response(self, template, **context):
-    renderer = jinja2.get_jinja2(app=self.app)
-    rendered_value = renderer.render_template(template, **context)
-    self.response.write(rendered_value)
-
-  @decorator.oauth_aware
-  def get(self):
-    if decorator.has_credentials():
-      result = service.tasks().list(tasklist='@default').execute(
-          http=decorator.http())
-      tasks = result.get('items', [])
-      for task in tasks:
-        task['title_short'] = truncate(task['title'], 26)
-      self.render_response('index.html', tasks=tasks)
-    else:
-      url = decorator.authorize_url()
-      self.render_response('index.html', tasks=[], authorize_url=url)
+    @decorator.oauth_aware
+    def get(self):
+        if decorator.has_credentials():
+            result = (
+                service.tasks().list(tasklist="@default").execute(http=decorator.http())
+            )
+            tasks = result.get("items", [])
+            for task in tasks:
+                task["title_short"] = truncate(task["title"], 26)
+            self.render_response("index.html", tasks=tasks)
+        else:
+            url = decorator.authorize_url()
+            self.render_response("index.html", tasks=[], authorize_url=url)
 
 
 def truncate(s, l):
-  return s[:l] + '...' if len(s) > l else s
+    return s[:l] + "..." if len(s) > l else s
 
 
-application = webapp2.WSGIApplication([
-    ('/', MainHandler),
-    (decorator.callback_path, decorator.callback_handler()),
-    ], debug=True)
+application = webapp2.WSGIApplication(
+    [
+        ("/", MainHandler),
+        (decorator.callback_path, decorator.callback_handler()),
+    ],
+    debug=True,
+)
