@@ -12,12 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from enum import IntEnum
-import numpy as np
-import pandas as pd
 import pathlib
 
 from changesummary import ChangeType
+import numpy as np
+import pandas as pd
 
 SCRIPTS_DIR = pathlib.Path(__file__).parent.resolve()
 CHANGE_SUMMARY_DIR = SCRIPTS_DIR / "temp"
@@ -48,7 +47,9 @@ class BuildPrBody:
         sha = None
         api_link = ""
 
-        file_path = pathlib.Path(self._change_summary_directory) / "{0}.sha".format(name)
+        file_path = pathlib.Path(self._change_summary_directory) / "{0}.sha".format(
+            name
+        )
         if file_path.is_file():
             with open(file_path, "r") as f:
                 sha = f.readline().rstrip()
