@@ -22,41 +22,42 @@ from __future__ import absolute_import
 
 __author__ = "jcgregorio@google.com (Joe Gregorio)"
 
+import io
 from io import FileIO
 
 # Do not remove the httplib2 import
 import json
-import httplib2
-import io
 import logging
-import mock
 import os
-import unittest
-import urllib
 import random
 import socket
 import ssl
 import time
+import unittest
+from unittest import mock
+import urllib
+
+import httplib2
+from oauth2client.client import Credentials
 
 from googleapiclient.discovery import build
-from googleapiclient.errors import BatchError
-from googleapiclient.errors import HttpError
-from googleapiclient.errors import InvalidChunkSizeError
-from googleapiclient.http import build_http
-from googleapiclient.http import BatchHttpRequest
-from googleapiclient.http import HttpMock
-from googleapiclient.http import HttpMockSequence
-from googleapiclient.http import HttpRequest
-from googleapiclient.http import MAX_URI_LENGTH
-from googleapiclient.http import MediaFileUpload
-from googleapiclient.http import MediaInMemoryUpload
-from googleapiclient.http import MediaIoBaseDownload
-from googleapiclient.http import MediaIoBaseUpload
-from googleapiclient.http import MediaUpload
-from googleapiclient.http import _StreamSlice
-from googleapiclient.http import set_user_agent
+from googleapiclient.errors import BatchError, HttpError, InvalidChunkSizeError
+from googleapiclient.http import (
+    MAX_URI_LENGTH,
+    BatchHttpRequest,
+    HttpMock,
+    HttpMockSequence,
+    HttpRequest,
+    MediaFileUpload,
+    MediaInMemoryUpload,
+    MediaIoBaseDownload,
+    MediaIoBaseUpload,
+    MediaUpload,
+    _StreamSlice,
+    build_http,
+    set_user_agent,
+)
 from googleapiclient.model import JsonModel
-from oauth2client.client import Credentials
 
 
 class MockCredentials(Credentials):
@@ -327,7 +328,6 @@ class TestMediaIoBaseUpload(unittest.TestCase):
         except NotImplementedError:
             pass
 
-
     def test_media_io_base_upload_from_bytes(self):
         f = open(datafile("small.png"), "rb")
         fd = io.BytesIO(f.read())
@@ -383,8 +383,8 @@ class TestMediaIoBaseUpload(unittest.TestCase):
         )
 
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/upload/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/upload/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http, model.response, uri, method=method, headers={}, resumable=upload
         )
@@ -407,8 +407,8 @@ class TestMediaIoBaseUpload(unittest.TestCase):
         )
 
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/upload/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/upload/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http, model.response, uri, method=method, headers={}, resumable=upload
         )
@@ -420,7 +420,6 @@ class TestMediaIoBaseUpload(unittest.TestCase):
             request.execute(num_retries=3)
         request._sleep.assert_not_called()
 
-
     def test_media_io_base_empty_file(self):
         fd = io.BytesIO()
         upload = MediaIoBaseUpload(
@@ -429,14 +428,26 @@ class TestMediaIoBaseUpload(unittest.TestCase):
 
         http = HttpMockSequence(
             [
-                ({"status": "200", "location": "https://www.googleapis.com/someapi/v1/upload?foo=bar"}, "{}"),
-                ({"status": "200", "location": "https://www.googleapis.com/someapi/v1/upload?foo=bar"}, "{}")
+                (
+                    {
+                        "status": "200",
+                        "location": "https://www.googleapis.com/someapi/v1/upload?foo=bar",
+                    },
+                    "{}",
+                ),
+                (
+                    {
+                        "status": "200",
+                        "location": "https://www.googleapis.com/someapi/v1/upload?foo=bar",
+                    },
+                    "{}",
+                ),
             ]
         )
 
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/upload/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/upload/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http, model.response, uri, method=method, headers={}, resumable=upload
         )
@@ -912,14 +923,14 @@ class TestHttpRequest(unittest.TestCase):
     def test_unicode(self):
         http = HttpMock(datafile("zoo.json"), headers={"status": "200"})
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
         request.execute()
@@ -931,8 +942,8 @@ class TestHttpRequest(unittest.TestCase):
     def test_empty_content_type(self):
         """Test for #284"""
         http = HttpMock(None, headers={"status": 200})
-        uri = u"https://www.googleapis.com/someapi/v1/upload/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/upload/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http, _postproc_none, uri, method=method, headers={"content-type": ""}
         )
@@ -944,7 +955,7 @@ class TestHttpRequest(unittest.TestCase):
         request = HttpRequest(
             HttpMockWithNonRetriableErrors(1, {"status": "200"}, '{"foo": "bar"}'),
             model.response,
-            u"https://www.example.com/json_api_endpoint",
+            "https://www.example.com/json_api_endpoint",
         )
         request._sleep = lambda _x: 0  # do nothing
         request._rand = lambda: 10
@@ -956,12 +967,12 @@ class TestHttpRequest(unittest.TestCase):
         request = HttpRequest(
             HttpMockWithErrors(5, {"status": "200"}, '{"foo": "bar"}'),
             model.response,
-            u"https://www.example.com/json_api_endpoint",
+            "https://www.example.com/json_api_endpoint",
         )
         request._sleep = lambda _x: 0  # do nothing
         request._rand = lambda: 10
         response = request.execute(num_retries=5)
-        self.assertEqual({u"foo": u"bar"}, response)
+        self.assertEqual({"foo": "bar"}, response)
 
     def test_retry_connection_errors_resumable(self):
         with open(datafile("small.png"), "rb") as small_png_file:
@@ -976,34 +987,38 @@ class TestHttpRequest(unittest.TestCase):
                 5, {"status": "200", "location": "location"}, '{"foo": "bar"}'
             ),
             model.response,
-            u"https://www.example.com/file_upload",
+            "https://www.example.com/file_upload",
             method="POST",
             resumable=upload,
         )
         request._sleep = lambda _x: 0  # do nothing
         request._rand = lambda: 10
         response = request.execute(num_retries=5)
-        self.assertEqual({u"foo": u"bar"}, response)
+        self.assertEqual({"foo": "bar"}, response)
 
     def test_retry(self):
         num_retries = 6
         resp_seq = [({"status": "500"}, "")] * (num_retries - 4)
         resp_seq.append(({"status": "403"}, RATE_LIMIT_EXCEEDED_RESPONSE))
-        resp_seq.append(({"status": "403"}, USER_RATE_LIMIT_EXCEEDED_RESPONSE_NO_STATUS))
-        resp_seq.append(({"status": "403"}, USER_RATE_LIMIT_EXCEEDED_RESPONSE_WITH_STATUS))
+        resp_seq.append(
+            ({"status": "403"}, USER_RATE_LIMIT_EXCEEDED_RESPONSE_NO_STATUS)
+        )
+        resp_seq.append(
+            ({"status": "403"}, USER_RATE_LIMIT_EXCEEDED_RESPONSE_WITH_STATUS)
+        )
         resp_seq.append(({"status": "429"}, ""))
         resp_seq.append(({"status": "200"}, "{}"))
 
         http = HttpMockSequence(resp_seq)
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1023,14 +1038,14 @@ class TestHttpRequest(unittest.TestCase):
 
         http = HttpMockSequence(resp_seq)
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1045,14 +1060,14 @@ class TestHttpRequest(unittest.TestCase):
     def test_no_retry_fails_fast(self):
         http = HttpMockSequence([({"status": "500"}, ""), ({"status": "200"}, "{}")])
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1068,14 +1083,14 @@ class TestHttpRequest(unittest.TestCase):
             [({"status": "403"}, NOT_CONFIGURED_RESPONSE), ({"status": "200"}, "{}")]
         )
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1089,14 +1104,14 @@ class TestHttpRequest(unittest.TestCase):
     def test_no_retry_403_fails_fast(self):
         http = HttpMockSequence([({"status": "403"}, ""), ({"status": "200"}, "{}")])
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1110,14 +1125,14 @@ class TestHttpRequest(unittest.TestCase):
     def test_no_retry_401_fails_fast(self):
         http = HttpMockSequence([({"status": "401"}, ""), ({"status": "200"}, "{}")])
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1136,14 +1151,14 @@ class TestHttpRequest(unittest.TestCase):
             ]
         )
         model = JsonModel()
-        uri = u"https://www.googleapis.com/someapi/v1/collection/?foo=bar"
-        method = u"POST"
+        uri = "https://www.googleapis.com/someapi/v1/collection/?foo=bar"
+        method = "POST"
         request = HttpRequest(
             http,
             model.response,
             uri,
             method=method,
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
         )
 
@@ -1198,7 +1213,7 @@ class TestBatch(unittest.TestCase):
             None,
             "https://www.googleapis.com/someapi/v1/collection/?foo=bar",
             method="POST",
-            body=u"{}",
+            body="{}",
             headers={"content-type": "application/json"},
             methodId=None,
             resumable=None,
@@ -1531,7 +1546,7 @@ class TestBatch(unittest.TestCase):
         self.assertEqual(
             "Authorization Required", callbacks.exceptions["1"].resp.reason
         )
-        self.assertEqual({u"baz": u"qux"}, callbacks.responses["2"])
+        self.assertEqual({"baz": "qux"}, callbacks.responses["2"])
         self.assertEqual(None, callbacks.exceptions["2"])
 
     def test_execute_global_callback(self):
