@@ -22,17 +22,14 @@ from __future__ import absolute_import
 
 __author__ = "jcgregorio@google.com (Joe Gregorio)"
 
-import httplib2
 import os
 import unittest
 
-from googleapiclient.errors import HttpError
-from googleapiclient.errors import UnexpectedBodyError
-from googleapiclient.errors import UnexpectedMethodError
-from googleapiclient.discovery import build
-from googleapiclient.http import RequestMockBuilder
-from googleapiclient.http import HttpMock
+import httplib2
 
+from googleapiclient.discovery import build
+from googleapiclient.errors import HttpError, UnexpectedBodyError, UnexpectedMethodError
+from googleapiclient.http import HttpMock, RequestMockBuilder
 
 DATA_DIR = os.path.join(os.path.dirname(__file__), "data")
 
@@ -48,7 +45,13 @@ class Mocks(unittest.TestCase):
 
     def test_default_response(self):
         requestBuilder = RequestMockBuilder({})
-        plus = build("plus", "v1", http=self.http, requestBuilder=requestBuilder, static_discovery=False)
+        plus = build(
+            "plus",
+            "v1",
+            http=self.http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
         activity = plus.activities().get(activityId="tag:blah").execute()
         self.assertEqual({}, activity)
 
@@ -56,7 +59,13 @@ class Mocks(unittest.TestCase):
         requestBuilder = RequestMockBuilder(
             {"plus.activities.get": (None, '{"foo": "bar"}')}
         )
-        plus = build("plus", "v1", http=self.http, requestBuilder=requestBuilder, static_discovery=False)
+        plus = build(
+            "plus",
+            "v1",
+            http=self.http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         activity = plus.activities().get(activityId="tag:blah").execute()
         self.assertEqual({"foo": "bar"}, activity)
@@ -64,7 +73,13 @@ class Mocks(unittest.TestCase):
     def test_unexpected_call(self):
         requestBuilder = RequestMockBuilder({}, check_unexpected=True)
 
-        plus = build("plus", "v1", http=self.http, requestBuilder=requestBuilder, static_discovery=False)
+        plus = build(
+            "plus",
+            "v1",
+            http=self.http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         try:
             plus.activities().get(activityId="tag:blah").execute()
@@ -76,7 +91,13 @@ class Mocks(unittest.TestCase):
         requestBuilder = RequestMockBuilder(
             {"zoo.animals.insert": (None, '{"data": {"foo": "bar"}}', None)}
         )
-        zoo = build("zoo", "v1", http=self.zoo_http, requestBuilder=requestBuilder, static_discovery=False)
+        zoo = build(
+            "zoo",
+            "v1",
+            http=self.zoo_http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         try:
             zoo.animals().insert(body="{}").execute()
@@ -88,7 +109,13 @@ class Mocks(unittest.TestCase):
         requestBuilder = RequestMockBuilder(
             {"zoo.animals.insert": (None, '{"data": {"foo": "bar"}}', "{}")}
         )
-        zoo = build("zoo", "v1", http=self.zoo_http, requestBuilder=requestBuilder, static_discovery=False)
+        zoo = build(
+            "zoo",
+            "v1",
+            http=self.zoo_http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         try:
             zoo.animals().insert(body="").execute()
@@ -106,7 +133,13 @@ class Mocks(unittest.TestCase):
                 )
             }
         )
-        zoo = build("zoo", "v1", http=self.zoo_http, requestBuilder=requestBuilder, static_discovery=False)
+        zoo = build(
+            "zoo",
+            "v1",
+            http=self.zoo_http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         try:
             zoo.animals().insert(body='{"data": {"foo": "blah"}}').execute()
@@ -124,7 +157,13 @@ class Mocks(unittest.TestCase):
                 )
             }
         )
-        zoo = build("zoo", "v1", http=self.zoo_http, requestBuilder=requestBuilder, static_discovery=False)
+        zoo = build(
+            "zoo",
+            "v1",
+            http=self.zoo_http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         activity = zoo.animals().insert(body={"data": {"foo": "bar"}}).execute()
         self.assertEqual({"foo": "bar"}, activity)
@@ -139,7 +178,13 @@ class Mocks(unittest.TestCase):
                 )
             }
         )
-        zoo = build("zoo", "v1", http=self.zoo_http, requestBuilder=requestBuilder, static_discovery=False)
+        zoo = build(
+            "zoo",
+            "v1",
+            http=self.zoo_http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         activity = zoo.animals().insert(body={"data": {"foo": "bar"}}).execute()
         self.assertEqual({"foo": "bar"}, activity)
@@ -149,7 +194,13 @@ class Mocks(unittest.TestCase):
         requestBuilder = RequestMockBuilder(
             {"plus.activities.list": (errorResponse, b"{}")}
         )
-        plus = build("plus", "v1", http=self.http, requestBuilder=requestBuilder, static_discovery=False)
+        plus = build(
+            "plus",
+            "v1",
+            http=self.http,
+            requestBuilder=requestBuilder,
+            static_discovery=False,
+        )
 
         try:
             activity = (
