@@ -25,9 +25,7 @@ __author__ = "jcgregorio@google.com (Joe Gregorio)"
 
 import unittest
 
-from googleapiclient.model import BaseModel
-from googleapiclient.model import makepatch
-
+from googleapiclient.model import BaseModel, makepatch
 
 TEST_CASES = [
     # (message, original, modified, expected)
@@ -76,15 +74,15 @@ class TestBaseModel(unittest.TestCase):
 
         test_cases = [
             ("hello", "world", "?hello=world"),
-            ("hello", u"world", "?hello=world"),
+            ("hello", "world", "?hello=world"),
             ("hello", "세계", "?hello=%EC%84%B8%EA%B3%84"),
-            ("hello", u"세계", "?hello=%EC%84%B8%EA%B3%84"),
+            ("hello", "세계", "?hello=%EC%84%B8%EA%B3%84"),
             ("hello", "こんにちは", "?hello=%E3%81%93%E3%82%93%E3%81%AB%E3%81%A1%E3%81%AF"),
-            ("hello", u"こんにちは", "?hello=%E3%81%93%E3%82%93%E3%81%AB%E3%81%A1%E3%81%AF"),
+            ("hello", "こんにちは", "?hello=%E3%81%93%E3%82%93%E3%81%AB%E3%81%A1%E3%81%AF"),
             ("hello", "你好", "?hello=%E4%BD%A0%E5%A5%BD"),
-            ("hello", u"你好", "?hello=%E4%BD%A0%E5%A5%BD"),
+            ("hello", "你好", "?hello=%E4%BD%A0%E5%A5%BD"),
             ("hello", "مرحبا", "?hello=%D9%85%D8%B1%D8%AD%D8%A8%D8%A7"),
-            ("hello", u"مرحبا", "?hello=%D9%85%D8%B1%D8%AD%D8%A8%D8%A7"),
+            ("hello", "مرحبا", "?hello=%D9%85%D8%B1%D8%AD%D8%A8%D8%A7"),
         ]
 
         for case in test_cases:
